test(evm): move peephole tests
