+++ crates/codegen/src/backend/evm/assembler.rs
@@ -933,108 +933,6 @@ mod tests {
     }
 
     #[test]
-    fn peephole_removes_push_zero_add() {
-        let mut asm = Assembler::new();
-
-        asm.emit_push(U256::from(42));
-        asm.emit_push(U256::ZERO);
-        asm.emit_op(op::ADD);
-        asm.emit_op(op::STOP);
-
-        let result = asm.assemble();
-
-        assert_eq!(result.bytecode, vec![0x60, 42, op::STOP]);
-    }
-
-    #[test]
-    fn peephole_cascades_after_rewrite() {
-        let mut asm = Assembler::new();
-
-        asm.emit_push(U256::from(42));
-        asm.emit_push(U256::ZERO);
-        asm.emit_op(op::ADD);
-        asm.emit_op(op::POP);
-
-        let result = asm.assemble();
-
-        assert!(result.bytecode.is_empty());
-    }
-
-    #[test]
-    fn peephole_resolves_labels_after_rewrites() {
-        let mut asm = Assembler::new();
-        let label = asm.new_label();
-
-        asm.emit_push(U256::from(42));
-        asm.emit_push(U256::ZERO);
-        asm.emit_op(op::ADD);
-        asm.define_label(label);
-        asm.emit_push_label(label);
-        asm.emit_op(op::JUMP);
-
-        let result = asm.assemble();
-
-        assert_eq!(result.label_offsets[&label], 2);
-        assert_eq!(result.bytecode, vec![0x60, 42, op::JUMPDEST, 0x60, 2, op::JUMP]);
-    }
-
-    #[test]
-    fn peephole_replaces_mul_zero_with_pop_zero() {
-        let mut asm = Assembler::new();
-
-        asm.emit_push(U256::from(42));
-        asm.emit_push(U256::ZERO);
-        asm.emit_op(op::MUL);
-        asm.emit_op(op::STOP);
-
-        let result = asm.assemble();
-
-        assert_eq!(result.bytecode, vec![op::PUSH0, op::STOP]);
-    }
-
-    #[test]
-    fn peephole_preserves_push_zero_sub() {
-        let mut asm = Assembler::new();
-
-        asm.emit_push(U256::from(42));
-        asm.emit_push(U256::ZERO);
-        asm.emit_op(op::SUB);
-        asm.emit_op(op::STOP);
-
-        let result = asm.assemble();
-
-        assert_eq!(result.bytecode, vec![0x60, 42, op::PUSH0, op::SUB, op::STOP]);
-    }
-
-    #[test]
-    fn peephole_rewrites_push_zero_eq() {
-        let mut asm = Assembler::new();
-
-        asm.emit_push(U256::from(42));
-        asm.emit_push(U256::ZERO);
-        asm.emit_op(op::EQ);
-        asm.emit_op(op::STOP);
-
-        let result = asm.assemble();
-
-        assert_eq!(result.bytecode, vec![0x60, 42, op::ISZERO, op::STOP]);
-    }
-
-    #[test]
-    fn peephole_preserves_push_one_div() {
-        let mut asm = Assembler::new();
-
-        asm.emit_push(U256::from(42));
-        asm.emit_push(U256::from(1));
-        asm.emit_op(op::DIV);
-        asm.emit_op(op::STOP);
-
-        let result = asm.assemble();
-
-        assert_eq!(result.bytecode, vec![0x60, 42, 0x60, 1, op::DIV, op::STOP]);
-    }
-
-    #[test]
     fn compact_full_word_all_ones_push() {
         let mut asm = size_optimized_assembler();
